fix: PR-Review-Feedback umgesetzt
- useLgsCountdown-Composable extrahiert (keine doppelte Logik mehr)
- Ferienwohnungen-Banner-Bild: LGS-Bild → Wohnküche Schöne Aussicht
- EN News-Seite: hardcodierte Strings → t() i18n-Aufrufe
- Hero z-index-Schichtkommentar hinzugefügt
